refactor: artist req 공통 vo 별도 클래스로 분리
